+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the Activity Bar Separator extension will be documented in this file.
 
+## [0.0.4] - 2025-05-16
+- Added support for multiple separators (up to 10)
+- Added configuration to set number of separators
+- Added command to easily change separator count
+- Improved extension structure and performance
+- Changed category to "Visualization"
+
 ## [0.0.3] - 2025-05-15
 - Added icon